signup_backend/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
python-dotenv==1.0.0
Flask-Cors==4.0.0
mysql-connector-python==8.1.0
Werkzeug==2.3.6
